+++ .github/workflows/deploy.yaml
@@ -4,6 +4,7 @@ on:
   push:
     branches:
       - main
+      - feat/mobileNav
 
 jobs:
   build:
