+++ components/layout/navbar/NavtreeRow.tsx
@@ -14,19 +14,19 @@ type NavTreeRowProps = {
   marginBottom: string;
 };
 
-export default function NavTreeRow({ segmentNode, highlight, marginBottom }: NavTreeRowProps) {
+export default function NavTreeLabel({ segmentNode, highlight, marginBottom }: NavTreeRowProps) {
   const href = getPath(segmentNode);
 
   if (highlight) {
-    return <HighlightedRow href={href} text={segmentNode.name} marginBottom={marginBottom} />;
+    return <HighlightedLabel href={href} text={segmentNode.name} marginBottom={marginBottom} />;
   } else if (segmentNode.isPage) {
-    return <LinkRow href={href} text={segmentNode.name} marginBottom={marginBottom} />;
+    return <LinkLabel href={href} text={segmentNode.name} marginBottom={marginBottom} />;
   } else {
-    return <TextRow text={segmentNode.name} marginBottom={marginBottom} />;
+    return <TextLabel text={segmentNode.name} marginBottom={marginBottom} />;
   }
 }
 
-function HighlightedRow({
+function HighlightedLabel({
   href,
   text,
   marginBottom,
@@ -48,7 +48,7 @@ function HighlightedRow({
   );
 }
 
-function LinkRow({
+function LinkLabel({
   href,
   text,
   marginBottom,
@@ -68,7 +68,7 @@ function LinkRow({
   );
 }
 
-function TextRow({ text, marginBottom }: { text: string; marginBottom: string }) {
+function TextLabel({ text, marginBottom }: { text: string; marginBottom: string }) {
   return (
     <p
       className="mb-6 block h-[1.0625rem] text-md font-medium leading-5 text-white"
